--- backend/controllers/hello-world.js
@@ -1,31 +1,0 @@
-export const getHelloWorld = (req, res) => {
-  res.status(200).json("Hello World!");
-};
-
-
-
-const bcrypt = require("bcrypt");
-const router = require("express").Router();
-
-router.post("/", async (req, res) => {
-  try {
-    const { error } = validate(req.body);
-    if (error)
-      return res.status(400).send({ message: error.details[0].message });
-
-    const user = await User.findOne({ email: req.body.email });
-    if (user)
-      return res.status(409).send({ message: "User with given email already exists!" });
-
-    const salt = await bcrypt.genSalt(Number(process.env.SALT));
-    const hashPassword = await bcrypt.hash(req.body.password, salt);
-
-    await new User({ ...req.body, password: hashPassword }).save();
-    res.status(201).send({ message: "User created succesfully!" })
-  } catch (error) {
-    console.log(error);
-    res.status(500).send({ message: "Internal server error!" })
-  }
-});
-
-module.exports = router;
