+++ backend/controllers/my-controller.js
@@ -1,0 +1,12 @@
+
+import bcrypt from 'bcrypt'
+
+
+
+export const getHelloWorld = (req, res) => {
+  res.status(200).json("Hello World!");
+};
+
+
+
+
