+++ backend/db/connection.js
@@ -1,8 +1,8 @@
 import mongoose from "mongoose";
 import "dotenv/config.js";
 
-// Define your connection URI here
-// const MONGODB_URI =
+
+const MONGODB_URI = "mongodb://localhost/users"
 
 mongoose.set("strictQuery", false);
 mongoose.set("returnOriginal", false); //for findByAndUpdate to return a reference to object at location
